00113: Add the microtime routines, add dcf clock support fixes
AUTHOR:	Sean Eric Fagan (sef@moria.cygnus.com) (microtime support)
AUTHOR:	Poul-Henning Kamp (phk@data.fls.dk) (dcfclock support)

	This patch makes the microtimer kernel routines work correctly
so that you can get more precise timestamps from gettimeofday().

	It also adds the necessary fixes to be able to install and
use the dcfclock driver.  The driver must be obtain from Poul-Henning Kamp
at this time, as it is still considered in test.  All that you need though
is the driver file itself, all the other changes have been incorporated into
the patchkit.  This should also increase the accuracy of the XTAL clock due
to Poul-Henning Kamp's other clock fixes.

NOTE:	The update for files.i386 is in a seperate patch so that it
could be combined with the new device support patch.